backend/requirements.txt
fastapi>=0.115.2,<1.0
uvicorn>=0.29.0
mangum>=0.17.0
motor>=3.4.0
pymongo>=4.7.2
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
httpx>=0.27.0
Pillow>=10.3.0
numpy>=1.26.4,<2.0
scikit-learn>=1.4.2
joblib>=1.4.2
python-dotenv>=1.0.1
